refactor main.py; save debug runs in a separate directory
